+++ lib/resque/kubernetes.rb
@@ -7,7 +7,6 @@ require "resque/kubernetes/dns_safe_random"
 require "resque/kubernetes/job"
 require "resque/kubernetes/jobs_manager"
 require "resque/kubernetes/version"
-require "resque/kubernetes/worker"
 
 module Resque
   # Run Resque Jobs as Kubernetes Jobs with autoscaling.
@@ -21,5 +20,3 @@ module Resque
     define_setting :max_workers, 10
   end
 end
-
-Resque::Worker.include Resque::Kubernetes::Worker